requirements.txt
argparse>=1.4.0  
requests>=2.28
PyYAML>=6.0
jsonpath-ng>=1.5.3
pytest>=7.0
requests-mock>=1.9